Split out blog post components
