--- freelance.md
@@ -1,36 +1,0 @@
-# Offerings
-
-## Responsive Web Design
-The web has gone mobile. Your sites need to be optimized for mobile phones, tablets and desktop screens alike. For this reason all of my sites are fully responsive - they adapt to different screen sizes so that your visitors always have the best experience.
-
-## E-Commerce
-Expand your business by selling products online. E-Commerce is here to stay and has fundamentally changed the way people shop. Small businesses are going global with new platforms like Shopify. Lets grow your business with an engaging online store.
-
-## Progressive Web Applications
-Your users are viewing your site on mobile browsers with spotty internet connections. They also expect fully functioning app like experiences in mobile browsers. I can help you build custom end to end web applications that feel native and work offline.
-
-# Small Business Stack
-
-# Principled Approach
-Offline experiences
-Mobile first
-Convenience
-Social
-
-# About Me
-
-
-
-
-
-
-
-
-## Responsive Web Design
-Your users are browsing the web on many different devices. Mobile phones, tablets, laptops etc. Responsive web design ensures that your website looks great on all devices so that your users have an amazing and consistent experience. All of my sites adapt to different screen sizes to best serve my your business.
-
-## E-Commerce
-If you haven't noticed, people love online shopping. Now, with platforms like Shopify, small businesses can have global reach by selling their products online. Whether you're just starting a business or taking your brick and mortar shop to the web we can grow your business together by creating an engaging e-commerce site.
-
-## Progressive Web Applications
-Users are expecting mobile sites to be able to do more and more. Luckily Progressive Web Applications solve this problem by introducing a set of best practices that embrace offline, mobile first web applications that feel like real mobile apps. I can deliver a custom end to end solution for your business that feels like its from the future.
