+++ projects/frontend/src/components/Blog/index.js
@@ -1,1 +1,6 @@
-export { default as PostMeta } from './meta';
+export { default as BottomBar } from './BottomBar';
+export { default as PostConclusion } from './Conclusion';
+export { default as PostInfo } from './Info';
+export { default as PostMeta } from './Meta';
+export { default as PostText } from './Text';
+export { default as ShareRow } from './Share';
